+++ lib/kubes/cli/apply.rb
@@ -3,7 +3,8 @@ class Kubes::CLI
     def run
       compile
       logger.info "Deploying kubes resources"
-      Kubes::Kubectl::Decider.new(:apply, @options).run
+      Kubes::Kubectl::Dispatcher.new(:apply, @options).run
+      Prune.new(@options.merge(yes: true, quiet: true)).run if Kubes.config.auto_prune # prune old secrets and config maps
     end
   end
 end