+++ lib/kubes/config.rb
@@ -11,31 +11,36 @@ module Kubes
     def defaults
       config = ActiveSupport::OrderedOptions.new
 
-      config.state = ActiveSupport::OrderedOptions.new
-      config.state.docker_image_path = "#{Kubes.root}/.kubes/state/docker_image.txt"
+      config.auto_prune = true
 
-      config.logger = Logger.new($stdout)
-      config.logger.level = ENV['KUBES_LOG_LEVEL'] || :info
+      config.builder = "docker" # IE: docker or gcloud
 
       # Auto-switching options
       config.kubectl = ActiveSupport::OrderedOptions.new
       config.kubectl.context = nil
       config.kubectl.context_keep = true # after switching context keep it
-      config.kubectl.exit_on_fail = nil # whether or not continue if the kubectl command fails
 
-      config.kubectl.exit_on_fail_for_apply  = true # whether or not continue if the kubectl apply command fails
-      config.kubectl.exit_on_fail_for_delete = false # whether or not continue if the kubectl delete command fails
-      # Note: not using config.kubectl.delete.exit_on_fail because delete a method internal to ActiveSupport::OrderedOptions
+      # whether or not continue if the kubectl command fails
+      config.kubectl.exit_on_fail = ActiveSupport::OrderedOptions.new
+      config.kubectl.exit_on_fail.apply  = true # whether or not continue if the kubectl apply command fails
+      config.kubectl.exit_on_fail.delete = false # whether or not continue if the kubectl delete command fails
+      # Note: delete is a internal method to ActiveSupport::OrderedOptions so will have to access it with ['...']
 
       config.kubectl.order = ActiveSupport::OrderedOptions.new
       config.kubectl.order.roles = role_order
       config.kubectl.order.kinds = kind_order
 
-      config.suffix_hash = true # append suffix has to ConfigMap and Secret
-
       config.repo = nil # expected to be set by .kubes/config.rb
 
-      config.builder = "docker" # IE: docker or gcloud
+      config.logger = Logger.new($stdout)
+      config.logger.level = ENV['KUBES_LOG_LEVEL'] || :info
+
+      config.skip = []
+
+      config.state = ActiveSupport::OrderedOptions.new
+      config.state.docker_image_path = "#{Kubes.root}/.kubes/state/docker_image.txt"
+
+      config.suffix_hash = true # append suffix hash to ConfigMap and Secret
 
       config
     end