+++ lib/kubes/kubectl/fetch/base.rb
@@ -2,6 +2,7 @@ require "json"
 
 module Kubes::Kubectl::Fetch
   class Base
+    extend Memoist
     include Kubes::Logging
     include Kubes::Util::Sh
 
@@ -9,16 +10,18 @@ module Kubes::Kubectl::Fetch
       @options = options
     end
 
-    def fetch_items
-      o = {
-        capture: true,
-        output: "json",
-        show_command: false,
-      }
-      kubectl = Kubes::Kubectl.new(:get, @options.merge(o)) # kubes get -f .kubes/output
-      resp = kubectl.run
-      data = JSON.load(resp)
+    def fetch(kind)
+      return [] unless namespace
+      data = Kubes::Kubectl.capture("get #{kind} -o json -n #{namespace}")
       data['items'] || [] # Note: When fetching only 1 resource, items is not part of structure
     end
+
+    def namespace
+      path = ".kubes/output/shared/namespace.yaml"
+      return unless File.exist?(path)
+      data = Kubes::Kubectl.capture("get -f #{path} -o json")
+      data['metadata']['name']
+    end
+    memoize :namespace
   end
 end