+++ lib/kubes/kubectl/fetch/deployment.rb
@@ -3,40 +3,39 @@ module Kubes::Kubectl::Fetch
     extend Memoist
 
     def metadata
-      deployment['metadata']
+      deployment['metadata'] if found
     end
 
     def spec
-      deployment['spec']
+      deployment['spec'] if found
     end
 
     def deployment
-      items = fetch_items
+      items = fetch(:deployment)
       # Not checking if deployment exists because kubes will error on `kubes get` from missing deployments already
       deployments = items.select { |i| i['kind'] == "Deployment" }
 
-      if deployments.size > 1 && !@options[:name]
+      if !@options[:deployment] && !@options[:pod] && deployments.size > 1
         names = deployments.map { |d| d['metadata']['name'] }
         logger.info <<~EOL
           INFO: More than one deployment found.
           Deployment names: #{names.join(', ')}
           Using #{names.first}
-          Note: You can specify the deployment to use with --name or -n
+          Note: You can specify the deployment to use with --deployment or -d
         EOL
       end
 
-      deployment = find_deployment(deployments)
-      unless deployment
-        logger.error "ERROR: No deployment found".color(:red)
-        exit 1
-      end
-      deployment
+      find_deployment(deployments)
     end
     memoize :deployment
 
+    def found
+      !!deployment
+    end
+
     def find_deployment(deployments)
-      if @options[:name]
-        deployments.find { |d| d['metadata']['name'] == @options[:name] }
+      if @options[:deployment]
+        deployments.find { |d| d['metadata']['name'] == @options[:deployment] }
       else
         deployments.first
       end